+++ src/main/java/blackjack/domain/card/Cards.java
@@ -1,7 +1,10 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import blackjack.domain.card.pattern.Denomination;
+import blackjack.domain.player.Gamer;
 
 public class Cards {
