+++ src/main/java/blackjack/domain/card/pattern/Denomination.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card.pattern;
 
 public enum Denomination {
