+++ src/test/java/blackjack/domain/DealerTest.java
@@ -6,6 +6,11 @@ import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.pattern.Denomination;
+import blackjack.domain.card.pattern.Suit;
+import blackjack.domain.player.Dealer;
+
 public class DealerTest {
 
 	@Test
