+++ src/test/java/blackjack/domain/DeckTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.card.Card;
+
 public class DeckTest {
 
 	private Deck deck;
