+++ src/test/java/blackjack/domain/DenominationTest.java
@@ -5,6 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.card.pattern.Denomination;
+
 public class DenominationTest {
 
 	@Test
