+++ src/test/java/blackjack/domain/ResultTest.java
@@ -5,6 +5,13 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.pattern.Denomination;
+import blackjack.domain.card.pattern.Suit;
+import blackjack.domain.player.Dealer;
+import blackjack.domain.player.Gamer;
+import blackjack.domain.player.Player;
+
 public class ResultTest {
 
 	@Test
